+++ src/identifier.cpp
@@ -1,0 +1,3 @@
+#include "identifier.hpp"
+
+uint32_t Identifier::NEXT_ID = 1;
